requirements.txt
requests
websockets
solana
base58
pandas
psycopg2-binary
python-dotenv